Send 16 button states instead of 8

+++ controller/include/Buttons.hpp
@@ -6,11 +6,11 @@ class Buttons {
 public:
     void initialize();
     bool update();
-    uint8_t state() { return m_state; }
+    uint16_t state() { return m_state; }
 
 private:
     Adafruit_MCP23017 m_mcp {};
 
-    uint8_t m_state {};
-    uint8_t m_last_state {};
+    uint16_t m_state {};
+    uint16_t m_last_state {};
 };
